+++ packages/sui-bundler/webpack.config.prod.js
@@ -8,22 +8,22 @@ const HtmlWebpackPlugin = require('html-webpack-plugin')
 const {WebpackManifestPlugin} = require('webpack-manifest-plugin')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 const InlineChunkHtmlPlugin = require('./shared/inline-chunk-html-plugin.js')
-const SpeedMeasurePlugin = require('speed-measure-webpack-plugin')
+
 const {
   when,
   cleanList,
   envVars,
   MAIN_ENTRY_POINT,
   config
-} = require('./shared/index')
-const minifyJs = require('./shared/minify-js')
-const minifyCss = require('./shared/minify-css')
-const definePlugin = require('./shared/define')
-const babelRules = require('./shared/module-rules-babel')
-const manifestLoaderRules = require('./shared/module-rules-manifest-loader')
-const {extractComments, sourceMap} = require('./shared/config')
-const {aliasFromConfig} = require('./shared/resolve-alias')
-const {resolveLoader} = require('./shared/resolve-loader')
+} = require('./shared/index.js')
+const {aliasFromConfig} = require('./shared/resolve-alias.js')
+const {extractComments, sourceMap} = require('./shared/config.js')
+const {resolveLoader} = require('./shared/resolve-loader.js')
+const babelRules = require('./shared/module-rules-babel.js')
+const definePlugin = require('./shared/define.js')
+const manifestLoaderRules = require('./shared/module-rules-manifest-loader.js')
+const minifyCss = require('./shared/minify-css.js')
+const minifyJs = require('./shared/minify-js.js')
 
 const PUBLIC_PATH = process.env.CDN || config.cdn || '/'
 
@@ -34,8 +34,6 @@ const filename = config.onlyHash
 const cssFileName = config.onlyHash
   ? '[contenthash:8].css'
   : '[name].[contenthash:8].css'
-
-const smp = new SpeedMeasurePlugin()
 
 /** @typedef {import('webpack').Configuration} WebpackConfig */
 
@@ -67,9 +65,15 @@ const webpackConfig = {
     minimizer: [minifyJs({extractComments, sourceMap}), minifyCss()].filter(
       Boolean
     ),
-    runtimeChunk: true
+    runtimeChunk: true,
+    splitChunks: {
+      chunks: 'all'
+    }
   },
   plugins: cleanList([
+    new webpack.ProvidePlugin({
+      process: 'process/browser'
+    }),
     new webpack.ids.HashedModuleIdsPlugin(),
     new webpack.EnvironmentPlugin(envVars(config.env)),
     definePlugin(),
@@ -133,4 +137,4 @@ const webpackConfig = {
   resolveLoader
 }
 
-module.exports = config.measure ? smp.wrap(webpackConfig) : webpackConfig
+module.exports = webpackConfig