+++ packages/sui-ssr/server/utils/index.js
@@ -1,10 +1,10 @@
 import path from 'path'
 import fs from 'fs'
-import config from '../config'
-import utilsFactory from './factory'
+import config from '../config.js'
+import utilsFactory from './factory.js'
 
 const {
-  buildRequestUrl,
+  getAssetsManifest,
   createStylesFor,
   hostFromReq,
   hrTimeToMs,
@@ -18,7 +18,7 @@ const {
 } = utilsFactory({path, fs, config})
 
 export {
-  buildRequestUrl,
+  getAssetsManifest,
   createStylesFor,
   hostFromReq,
   hrTimeToMs,